-serveis -adaptadors als serveis -AdapterFactory -pagament (amb una nova funcio a Representacio)

+++ Shows/src/com/Shows/Domain/Adapters/IConversorAdapter.java
@@ -5,5 +5,5 @@ import com.Shows.Domain.Model.Moneda;
 
 public interface IConversorAdapter {    
 	
-	public float convert(Moneda divisa, Moneda moneda);
+	public double convert(Moneda divisa, Moneda moneda);
 }
